created wp7 project and ported over
It's actually extremely annoying. The first project I created was a WP8
project, but since my phone hasn't updated to Windows Phone 7.8, when I
tried to deploy the app it told me that a Windows Phone 8 was not
detected. Thus, since my phone is still running WP7.1 I had to  build a
whole new project with different build settings. That was stupid.