saving videos to XML file!
Pretty tough problem. I create a .xml file, which I update each time the
app is closed and open for reading each time the app is opened. It's not
a fantastic solution, but it works for these purposes.

The XML object would be easily sent to a web server now.