saving and loading videos improving
Renamed the image files such that they have unique names and can be
loaded correctly. I should do the same with the video files and allow
them to be played. I think I can do that fairly quickly. I'm also still
having the image rotation issues which sucks, but I suppose we'll roll
with it.